[TWINFACES-458] feat: destory twin class link entitie and remove to separate entities
