+++ src/app/workspace/twins/[twinId]/views/twin-links.tsx
@@ -10,6 +10,7 @@ import { ColumnDef } from "@tanstack/table-core";
 import { useContext, useRef } from "react";
 import { toast } from "sonner";
 import { TwinContext } from "../twin-context";
+import { LinkResourceLink } from "@/entities/link";
 
 const colDefs: Record<
   keyof Pick<
@@ -32,7 +33,7 @@ const colDefs: Record<
     cell: ({ row: { original } }) =>
       original.link && (
         <div className="max-w-48 inline-flex">
-          <TwinClassLinkResourceLink data={original.link} withTooltip />
+          <LinkResourceLink data={original.link} withTooltip />
         </div>
       ),
   },
