+++ src/entities/link/components/index.ts
