+++ src/entities/link/components/resource-link/index.ts
