+++ src/entities/link/index.ts
@@ -1,2 +1,3 @@
 export * from "./api";
 export * from "./libs";
+export * from "./components";
