--- src/entities/twin-class-link/api/index.ts
@@ -1,1 +1,0 @@
-export * from "./api-service";
